requirements.txt
PyOpenGL>=3.1.6
PyOpenGL-accelerate>=3.1.6
glfw>=2.6.0
numpy>=1.24.0
